+++ core/models/__init__.py
@@ -1,4 +1,4 @@
 from .user import *
-from .year_tutor import *
+from .programme_tutor import *
 from .reviewer import *
 from .module import *
