+++ core/views/admin/__init__.py
@@ -1,4 +1,4 @@
 from .admin_module_views import *
 from .admin_user_views import *
-from .admin_year_tutor_views import *
+from .admin_programme_tutor_views import *
 from .admin_reviewer_views import *
